fix(Calculator) : 비정상적인 수익률 개선
int로 받아서 overflow 현상이 나타나는 버그를 해결한다

+++ src/main/java/lotto/Calculator.java
@@ -12,7 +12,7 @@ public class Calculator {
     }
 
     public double profit() {
-        int benefit = 0;
+        long benefit = 0;
 
         for (Prize prize : prizes) {
             benefit += prize.reward();
